[Behat][Checkout] Raw moving shipping steps
